WIP: moved more functions and added todo labels
