+++ src/neutralb1/analysis/pwa_tools.py
@@ -5,8 +5,6 @@ analyzing fit results
 """
 
 import cmath
-import itertools
-import re
 import warnings
 from typing import Dict, List, Literal, Optional, cast
 
@@ -154,6 +152,7 @@ class Plotter:
 
             # set non-existent columns to zero. Needed since the fit dataframe may use
             # a different waveset from the truth fit
+            # TODO: Move this to a preprocessing method
             cols_to_add = []
             for sum_type in self.coherent_sums:
                 for sum in self.coherent_sums[sum_type]:
@@ -1253,6 +1252,7 @@ class Plotter:
         """
         pass
 
+    # TODO: move to plotter, but consider if any components can be in statistics
     def bootstrap_matrix(
         self, fits: List[str] | str, columns: List[str], **kwargs
     ) -> None:
@@ -1731,6 +1731,7 @@ class Plotter:
 
         pass
 
+    # TODO: move to statistics
     def normality_test(
         self,
         columns: Optional[List[str]] = None,
@@ -1802,6 +1803,7 @@ class Plotter:
 
         pass
 
+    # TODO: move to preprocessing
     def _reset_truth_phases(self, df: pd.DataFrame, mass_bins: pd.Series) -> None:
         """Reset the phase differences of mass dependent truth csv
 
@@ -1861,6 +1863,7 @@ class Plotter:
             )
         pass
 
+    # TODO: move to preprocessing
     def _link_fits_to_dataframes(self) -> None:
 
         # Extract parent directory for each fit file
@@ -1928,241 +1931,3 @@ class Plotter:
             self.truth_df = pd.concat([self.truth_df, truth_df["fit_index"]], axis=1)
 
         pass
-
-
-# TODO: Move to utils
-def wrap_phases(obj: pd.DataFrame | pd.Series) -> None:
-    """Wrap phase differences to be from (-pi, pi] & convert from radians to degrees
-
-    Accepts either a pandas DataFrame or Series. For DataFrames, only phase difference
-    columns and their errors are wrapped. For Series, all values are wrapped.
-
-    Args:
-        obj (pd.DataFrame | pd.Series): DataFrame of fit results or Series of phase
-            difference values.
-    Raises:
-        ValueError: If obj is not a DataFrame or Series.
-
-    Returns:
-        None: Edits the df or series in place
-    """
-
-    # wraps phase (in radians) to -pi < x <= pi and convert to degrees
-    def wrap(phase):
-        return np.rad2deg(np.angle(np.exp(1j * phase)))
-
-    if isinstance(obj, pd.Series):
-        obj.update(obj.apply(wrap))
-    elif isinstance(obj, pd.DataFrame):
-        phase_diffs = get_phase_differences(obj)
-        for col in set(phase_diffs.values()):
-            obj[col] = obj[col].apply(wrap)
-            obj[f"{col}_err"] = obj[f"{col}_err"].apply(wrap)
-    else:
-        raise ValueError("Input must be a pandas DataFrame or Series.")
-
-    return
-
-
-# TODO: move to utils
-def get_phase_differences(df: pd.DataFrame) -> Dict[tuple, str]:
-    """Returns dict of all the phase difference columns in the dataframe
-
-    The keys are specifically like (eJPmL_1, eJPmL_2), and there is no way to
-    know how those will be ordered in the dataframe a priori. We avoid this by
-    creating keys for either ordering and setting both of their values to the order
-    found in the dataframe.
-
-    Args:
-        df (pd.DataFrame): dataframe of fit results loaded from csv
-    Returns:
-        dict: key = tuple of both possible combinations of every amplitude, val = the
-            phase difference combination found in the dataframe
-    """
-    phase_differences = {}
-
-    # get all possible combinations of eJPmL columns, and add their phase difference
-    # column name if it exists (handles reverse ordering i.e. p1ppS_p1pmS & p1pmS_p1ppS)
-    all_combos = list(itertools.combinations(get_coherent_sums(df)["eJPmL"], 2))
-    columns = df.columns.values.tolist()
-    for combo in all_combos:
-        name = "_".join(combo)
-        reverse_name = "_".join(reversed(combo))
-
-        if name in columns:
-            phase_differences[combo] = name
-            phase_differences[tuple(reversed(combo))] = name
-        if reverse_name in columns:
-            phase_differences[combo] = reverse_name
-            phase_differences[tuple(reversed(combo))] = reverse_name
-
-    return phase_differences
-
-
-# TODO: move to utils
-def human_format(num: float) -> str:
-    """Converts orders of magnitude to letters i.e. 12369 -> 12.4K
-
-    Args:
-        num (float): positive floating point number
-
-    Returns:
-        str: first 3 significant digits with a character denoting its magnitude
-    """
-    num = float(f"{num:.3g}")
-    magnitude = 0
-    while abs(num) >= 1000:
-        magnitude += 1
-        num /= 1000.0
-    orders = ["", "K", "M", "B", "T"]
-    return f"{str(num).rstrip('0').rstrip('.')}{orders[magnitude]}"
-
-
-# TODO: move to physics
-def breakup_momentum(
-    parent_mass: float, daughter1_mass: float, daughter2_mass: float
-) -> float:
-    """Breakup momentum of a parent particle into two daughter particles
-
-    Pythonized version of https://github.com/JeffersonLab/halld_sim/blob/master/src/
-    libraries/AMPTOOLS_AMPS/breakupMomentum.cc. Returned value is independent of which
-    particle is considered daughter1 or daughter2. Take care to units are consistent
-    between masses.
-
-    Args:
-        parent_mass (float): mass of parent particle
-        daughter1_mass (float): mass of one of the daughter particles
-        daughter2_mass (float): mass of the other daughter particle
-
-    Returns:
-        float: Breakup momentum of the parent particle.
-    """
-    return np.sqrt(
-        np.abs(
-            np.power(parent_mass, 4)
-            + np.power(daughter1_mass, 4)
-            + np.power(daughter2_mass, 4)
-            - 2.0 * np.square(parent_mass) * np.square(daughter1_mass)
-            - 2.0 * np.square(parent_mass) * np.square(daughter2_mass)
-            - 2.0 * np.square(daughter1_mass) * np.square(daughter2_mass)
-        )
-    ) / (2.0 * parent_mass)
-
-
-# TODO: move to physics
-def barrier_factor(
-    parent_mass: float, daughter1_mass: float, daughter2_mass: float, l: int
-) -> float:
-    """Blatt-Weisskopf barrier factor for a given parent-daughter system
-
-    Pythonized version of https://github.com/JeffersonLab/halld_sim/blob/master/src/
-    libraries/AMPTOOLS_AMPS/barrierFactor.cc. This parameterization often suppresses
-    values near threshold, and amplifies values at higher masses, especially for large
-    l values.
-
-    Args:
-        parent_mass (float): mass of the parent particle
-        daughter1_mass (float): mass of the first daughter particle
-        daughter2_mass (float): mass of the second daughter particle
-        l (int): orbital angular momentum quantum number of the parent particle
-
-    Returns:
-        float: barrier factor
-    """
-
-    q = np.abs(breakup_momentum(parent_mass, daughter1_mass, daughter2_mass))
-    z = np.square(q) / np.square(0.1973)
-
-    match l:
-        case 0:
-            barrier = 1.0
-        case 1:
-            barrier = (2.0 * z) / (z + 1.0)
-        case 2:
-            barrier = (13.0 * np.square(z)) / (np.square(z - 3.0) + 9.0 * z)
-        case 3:
-            barrier = (277.0 * np.power(z, 3)) / (
-                z * np.square(z - 15.0) + 9.0 * np.square(2.0 * z - 5.0)
-            )
-        case 4:
-            barrier = (12746.0 * np.power(z, 4)) / (
-                np.square((np.square(z) - 45.0 * z + 105.0))
-                + 25.0 * z * np.square(2.0 * z - 21.0)
-            )
-        case _:
-            barrier = 0.0
-
-    return np.sqrt(barrier)
-
-
-# TODO: move to physics
-def breit_wigner(
-    mass: float,
-    bw_mass: float,
-    bw_width: float,
-    bw_l: int,
-    daughter1_mass: float = 0.1349768,
-    daughter2_mass: float = 0.78266,
-) -> complex:
-    """Halld_sim parameterization of the breit wigner function
-
-    To avoid discrepancies, this function copies the parameterization of the breit
-    wigner function found in https://github.com/JeffersonLab/halld_sim/blob/master/src/
-    libraries/AMPTOOLS_AMPS/BreitWigner.cc.
-
-    NOTE: all masses / widths must be in GeV. Daughter particle masses are typically
-    obtained from event 4-vectors, but since they're not available post-fit, we
-    approximate them by constraining their mass to the pdg value
-
-    Args:
-        mass (float): mass value to evaluate breit wigner function at. If using a mass
-            bin, approximate by passing the center of the bin
-        bw_mass (float): breit wigner central mass
-        bw_width (float): breit wigner width
-        bw_l (int): orbital angular momentum of the breit wigner i.e. rho->(omega,pi0)
-            is in the P-wave, so bw_l = 1
-        daughter1_mass (float, optional): mass of 1st daughter particle.
-            Defaults to 0.1349768 for the pi0 mass
-        daughter2_mass (float, optional): mass of 2nd daughter particle.
-            Defaults to 0.78266 for the omega mass
-
-    Returns:
-        complex: value of the breit wigner at the mass value
-    """
-
-    q0 = np.abs(breakup_momentum(bw_mass, daughter1_mass, daughter2_mass))
-    q = np.abs(breakup_momentum(mass, daughter1_mass, daughter2_mass))
-
-    F0 = barrier_factor(bw_mass, daughter1_mass, daughter2_mass, bw_l)
-    F = barrier_factor(mass, daughter1_mass, daughter2_mass, bw_l)
-
-    width = bw_width * (bw_mass / mass) * (q / q0) * np.square(F / F0)
-
-    numerator = complex(np.sqrt((bw_mass * bw_width) / np.pi), 0.0)
-    denominator = complex(np.square(bw_mass) - np.square(mass), -1.0 * bw_mass * width)
-
-    return F * numerator / denominator
-
-
-# TODO: Move to utils
-def char_to_int(char: str) -> int:
-    """Convert m-projection or L angular momenta character to integer value
-
-    Args:
-        char (str): single character. lowercase assumes m-projection and uppercase
-            assumes it's an L value
-    Returns:
-        int: integer that char is representing
-    """
-    d = {
-        "n": -2,
-        "m": -1,
-        "0": 0,
-        "p": +1,
-        "q": +2,
-        "S": 0,
-        "P": 1,
-        "D": 2,
-        "F": 3,
-    }
-    return d[char]
